requirement.txt
BeautifulSoup==3.2.1
Django==1.6.2
Markdown==2.4
Pillow==2.4.0
South==0.8.4
Unidecode==0.04.16
argparse==1.2.1
beautifulsoup4==4.3.2
django-appconf==0.6
django-compressor==1.3
django-filter==0.7
django-libsass==0.1
django-modelcluster==0.1
django-taggit==0.11.2
django-treebeard==2.0
djangorestframework==2.3.13
libsass==0.4.1
lxml==3.3.5
psycopg2==2.5.2
six==1.6.1
wsgiref==0.1.2
